Added section on installing the snap
Thanks for making VS Codium. Now the codium snap is available in the stable channel of the snap store, I added a section to the README.md so people can find it. I also added a badge at the top which shows what build is current in the snap store, and the build status.